Update Terminal++ API (#252)
The fundamental type of terminal data is now the unsigned byte. This puts it in line with Server++, Console++ and Telnet++. This aids interoperability.

Furthermore, the API is now based on continuations, meaning that there is no longer any returning of strings. Instead, byte sequences are passed forward to supplied functions. Initial measurements suggest this reduces memory allocations by about a third.

The API for the terminal class is now strictly stream-based, which removes the occasional error where the order of function calls when chaining together long strings involve cursor moves would cause strange effects.

Other structural changes are embedded in this: members are expressed consistently with an underscore suffix, event for public member variables. This is due to usability issues where the name frequently clashed with the type. In addition, relational operators are consistently expressed as member friends. This disables the potential for undesirable comparisons between unrelated types.