Terminals no longer write 8-bit control codes.
Due to a misunderstanding with the VT100 protocol, Terminal++ was sending
the code for activating 8-bit mode, but this should be sent from the
terminal and not the application.  A further change should allow the
reception of this code in order to switch the terminal mode.

For now, sending 8-bit control codes is removed, although received 8-bit
codes still parse correctly.

Fixes #255.